Move head into lib
