+++ lib/head.js
